requirements.txt
streamlit>=1.32.0
google-cloud-storage>=2.14.0
langchain>=0.1.12
langchain-core>=0.1.40
langchain-community>=0.0.29
langchain-text-splitters>=0.0.1
langchain-milvus>=0.1.5
langchain-groq>=0.1.6
groq>=0.9.0
sentence-transformers>=2.5.1
pymilvus>=2.4.0
duckduckgo-search>=5.3.0
pypdf>=4.1.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
tiktoken>=0.6.0
unstructured>=0.12.5
sentence-transformers